goodix: fix incompatible redeclaration of library function
drivers/input/touchscreen/tp_goodix_driver/test_core/board_opr_interface.h:51:12: warning: incompatible redeclaration of library function 'strtol' [-Wincompatible-library-redeclaration]
extern int strtol(const char *nptr, char **endptr, int base);
           ^
drivers/input/touchscreen/tp_goodix_driver/test_core/board_opr_interface.h:51:12: note: 'strtol' is a builtin with type 'long (const char *, char **, int)'

+++ drivers/input/touchscreen/tp_goodix_driver/test_core/board_opr_interface.h
@@ -48,7 +48,7 @@ extern "C" {
 #define EOF -1
 /*********************************str opr**********************************/
 extern char *strdup(const char *s);
-extern int strtol(const char *nptr, char **endptr, int base);
+extern int long strtol (const char *nptr, char **endptr, int base);
 
 /*********************************mem opr**********************************/
 extern void *malloc(size_t size);
